requirements.txt
av==9.2.0
click==8.1.3
cycler==0.11.0
fonttools==4.33.3
imageio==2.19.2
imageio-ffmpeg==0.4.7
imutils==0.5.4
kiwisolver==1.4.2
matplotlib==3.5.2
networkx==2.8.2
numpy==1.22.4
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyqt5-tools==5.15.4.3.2
PySide6==6.2.4
pytesseract==0.3.9
python-dateutil==2.8.2
python-dotenv==0.20.0
PyWavelets==1.3.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
scikit-image==0.19.2
scipy==1.8.1
shiboken6==6.2.4
six==1.16.0
tifffile==2022.5.4
tkVideo==0.1
tkvideoplayer==2.0.0
